requirements.txt
yfinance
python-telegram-bot
requests
schedule